extracting animation-logic from the useMediaSwitcher-hook to a customized hook
