+++ src/components/screens/home/components/Home-tv-show.spec.tsx
@@ -7,8 +7,8 @@ import * as SchemaTypes from '@schema-types';
 import * as TRANSLATIONS from '@i18n/tags';
 import { Routes } from '@routes/routes';
 
+import { SWITCH_ANIMATION_DURATION_MS } from './media-switcher/useMediaSwitcherAnimation';
 import timeTravel, { setupTimeTravel } from '../../../../../__mocks__/timeTravel';
-import { SWITCH_ANIMATION_DURATION_MS } from './media-switcher/useMediaSwitcher';
 import AutoMockProvider from '../../../../../__mocks__/AutoMockedProvider';
 import MockedNavigation from '../../../../../__mocks__/MockedNavigator';
 import { TRANSITIONING_DURATION } from './hooks/useHome';
