+++ src/components/screens/home/components/media-switcher/MediaSwitcher.spec.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from 'styled-components';
 import { dark as theme } from '@styles/themes/dark';
 
 import timeTravel, { setupTimeTravel } from '../../../../../../__mocks__/timeTravel';
-import { SWITCH_ANIMATION_DURATION_MS } from './useMediaSwitcher';
+import { SWITCH_ANIMATION_DURATION_MS } from './useMediaSwitcherAnimation';
 import MediaSwitcher from './MediaSwitcher';
 
 const ACTIVE_TEXT_RGBA = theme.colors.buttonText;
